página "minha conta" conta criada, correções e melhorias nas outras páginas
