+++ web/.eslintrc.js
@@ -1,0 +1,18 @@
+module.exports = {
+  env: {
+    browser: true,
+    es2021: true,
+  },
+  extends: ['plugin:react/recommended', 'airbnb'],
+  overrides: [],
+  parserOptions: {
+    ecmaVersion: 'latest',
+    sourceType: 'module',
+  },
+  plugins: ['react'],
+  rules: {
+    'import/no-named-as-default': 0,
+    'object-curly-newline': 0,
+    'react/prop-types': 0,
+  },
+};
