+++ web/src/components/CustomTable/Pagination.jsx
@@ -1,4 +1,6 @@
-import Form from "react-bootstrap/Form";
+import React from 'react';
+import Form from 'react-bootstrap/Form';
+import { CaretRightFill, CaretLeftFill } from 'react-bootstrap-icons';
 
 export default function Pagination({
   pageSize = 10,
@@ -8,7 +10,9 @@ export default function Pagination({
   total = 10,
 }) {
   const startPage = pageSize * (currentPage - 1) + 1;
-  const endPage = pageSize * currentPage;
+  const _endPage = pageSize * currentPage;
+  const endPage = _endPage > total ? total : _endPage;
+  const currentShowingDescription = `${startPage}-${endPage} de ${total}`;
 
   return (
     <div>
@@ -25,10 +29,9 @@ export default function Pagination({
         <option value="50">50</option>
         <option value="100">100</option>
       </Form.Select>
-      <p className="d-inline-block mx-1">{`${startPage}-${
-        endPage > total ? total : endPage
-      } de ${total}`}</p>
-
+      <p className="d-inline-block mx-1">{currentShowingDescription}</p>
+      <CaretLeftFill />
+      <CaretRightFill />
       {/* TODO: implement page navigation */}
     </div>
   );
