+++ web/src/pages/Dashboard/index.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Container, Row } from 'react-bootstrap';
+
+import './styles.css';
+
+function Dashboard() {
+  return (
+    <div className="dashboard-page">
+      <Container>
+        <Row></Row>
+      </Container>
+    </div>
+  );
+}
+
+export default Dashboard;
