+++ web/src/pages/MinhaConta/styles.css
@@ -1,0 +1,9 @@
+.minha-conta-page .divider {
+  border-right: 1px lightgray solid;
+}
+
+.minha-conta-page .readonly input:disabled,
+.minha-conta-page .readonly select:disabled {
+  background: inherit;
+  border: none;
+}
